Moved remote interfaces to definitions
