+++ src/definitions/RemoteConnection.ts
